third_party/WebKit/LayoutTests/media/media-can-play-mpeg-audio.html

<!DOCTYPE html>
<title>Test HTMLMediaElement's "canPlayType" method with unsupported audio mpeg MIME type.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<audio></audio>
<script>
test(function() {
    var audio = document.querySelector("audio");
    // Currently no platform supports the below formats.
    assert_equals(audio.canPlayType("audio/x-mpeg"), "");
});
</script>
